Create reusable workflow for our private repos
To reduce the effort of keepnig these two workflows in sync:

https://github.com/guardian/scala-steward-private-repos/blob/main/.github/workflows/scala-steward.yml
https://github.com/guardian/scala-steward-public-repos/blob/main/.github/workflows/scala-steward.yml

...we can let the public repo have a reusable workflow, and let the
private repo call it.

https://docs.github.com/en/actions/using-workflows/reusing-workflows

We forgot to add repo-checkout (like https://github.com/guardian/scala-steward-public-repos/pull/4)
to `.github/workflows/scala-steward.yml` in the private repo when doing
https://github.com/guardian/scala-steward-private-repos/pull/2,
so subsequent executions failed:

https://github.com/guardian/scala-steward-private-repos/actions/runs/2720202150

The error message was:

> The path indicated in `repos-file` (REPOSITORIES.md) does not exist

...`REPOSITORIES.md` was missing because the workflow didn't do
a repo checkout - originally it hadn't needed it.