Add CODEOWNERS file for Scala Guild
This ensures that the Scala Guild will automatically be requested for review
when someone opens a pull request on this repo.

https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners

Note that the team-by-team Grafana reports produced by DevX tend not to use
the information in the CODEOWNERS, the determination of which team is responsible
for a repository or security compliance issue tends to come down to one of these
two factors:

* For a repo: which GitHub teams have Admin access to the repo
* For an AWS security issue: which GitHub team owns the account

For instance, this report on dependency vulnerabilities on repos, which is
filtered for the Scala Guild, is showing repos for which the Scala Guild have
admin access:

https://metrics.gutools.co.uk/d/fdib3p8l85jwgd/dependency-vulnerabilities-by-team?orgId=1&var-repo_owner=scala-guild

It's a shame that we need to define ownership of Scala Guild repos in two
different ways (GitHub Team admin access & CODEOWNERS file) but I guess we
can live with it!