Remove "Non-AWS" group (#103)
The `pullRequests.grouping` field groups multiple dependency updates
into single PRs. We've found that in some projects the grouped PRs
generated by this grouping config are large and complex, and therefore
hard to review. We'd like smaller PRs with fewer changes.

We've tried to override this config at the individual project level,
but scala-steward bypasses these overrides and continues to use the
central config defined here. We're therefore removing some of these
defaults, starting with the "Non-AWS" group as it tends to create the
largest and most disparate PRs.

We believe that consuming projects should be able to add new groups,
they just can't remove existing ones, so any project that wants to
retain this group can add it to their individual config. We will test
this to make sure.